**ILP Solver - DEMO**
The demo version is going to be shared over github since on eclipse the
real one doesn't pass compilation. The function and the module has the
exact name, but the solution uses backtracking and not ILP. When we
compile on Nova we will simply change the ILPSolver.c file and copy the
suitable makefile to the directory.